defibrillator system move into shared
