+++ Content.Client/Medical/DefibrillatorSystem.cs
@@ -1,0 +1,10 @@
+using Content.Shared.Medical;
+
+namespace Content.Client.Medical;
+
+/// <summary>
+/// This handles client interactions and logic relating to <see cref="DefibrillatorComponent"/>
+/// </summary>
+public sealed class DefibrillatorSystem : SharedDefibrillatorSystem
+{
+}
